+++ recipe/commands/project.py
@@ -65,10 +65,6 @@ class ProjectCommand(Command):
 
             cookiecutter(temp_work_dir, no_input=True, output_dir=args.out)
 
-            self.logger.info('Execute POST script.')
-            output_project = os.path.join(args.out, project_slug)
-            self._post_generate(temp_work_dir, output_project=output_project)
-
             if self.options.deploy:
                 self.logger.info('Create jenkins CI jobs.')
                 jenkins = self.config.get_tuple('jenkins', 'url', 'user', 'password')
@@ -83,7 +79,7 @@ class ProjectCommand(Command):
         except Exception as e:
             self.logger.exception(e)
             self.logger.info("Create project %s failure.", project_slug)
-            sys.exit(3)
+            raise e
 
     def _post_generate(self, temp_work_dir, output_project=None):
         """ convert CRLF to LF line separator