+++ templates/newkit.app/{{cookiecutter.project_slug}}/app/demo/app.js
@@ -1,0 +1,6 @@
+angular.module('{{cookiecutter.project_slug}}')
+    .controller('DemoCtrl', [
+        "$scope", function ($scope) {
+
+        }
+    ]);