+++ templates/newkit.app/{{cookiecutter.project_slug}}/l18n/en-us.js
@@ -1,0 +1,5 @@
+NEG.namespace('resources.{{cookiecutter.project_slug}}.us', {
+    common: {
+        title: "Demo"
+    }
+});