+++ templates/newkit.app/{{cookiecutter.project_slug}}/l18n/zh-cn.js
@@ -1,0 +1,5 @@
+NEG.namespace('resources.{{cookiecutter.project_slug}}.cn', {
+    common: {
+        title: "例子"
+    }
+});