+++ templates/python.flask/{{cookiecutter.project_safe_name}}/config/default.py
@@ -1,0 +1,15 @@
+HTTP_HOST = '127.0.0.1'
+HTTP_PORT = 8080
+
+LOG = "/var/{{cookiecutter.project_slug}}"
+
+# WSGI Settings
+WSGI_LOG = 'default'
+
+# Flask-Log Settings
+LOG_LEVEL = 'debug'
+LOG_FILENAME = "logs/error.log"
+LOG_BACKUP_COUNT = 10
+LOG_MAX_BYTE = 1024 * 1024 * 10
+LOG_FORMATTER = '%(asctime)s - %(levelname)s - %(message)s'
+LOG_ENABLE_CONSOLE = True