+++ templates/python.flask/{{cookiecutter.project_safe_name}}/{{cookiecutter.project_slug}}/views/index.py
@@ -1,0 +1,8 @@
+from flask import render_template
+
+from {{cookiecutter.project_slug}} import bp
+
+
+@bp.route("/", methods=['GET'])
+def index():
+    return render_template('index.html')