+++ templates/python.flask/{{cookiecutter.project_slug}}/config/prd.py
@@ -1,9 +1,0 @@
-# prd environment setting
-
-# WSGI Settings
-WSGI_LOG = None
-
-# Flask-Log Settings
-LOG_LEVEL = 'debug'
-LOG_FILENAME = "/var/{{cookiecutter.project_slug}}/error.log"
-LOG_ENABLE_CONSOLE = False