+++ templates/python.restful/{{cookiecutter.project_safe_name}}/config/gqc.py
@@ -1,0 +1,10 @@
+# gqc environment setting
+HTTP_HOST = ''
+
+# WSGI Settings
+WSGI_LOG = None
+
+# Flask-Log Settings
+LOG_LEVEL = 'debug'
+LOG_FILENAME = "/var/{{cookiecutter.project_slug}}/error.log"
+LOG_ENABLE_CONSOLE = False