+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change log
 
+## [0.0.6]
+
+### Added
+- Add Python.restful template
+- Add Doc template
+- Add doc template
+- Send email when occur unknown exception
+
+### Changed
+- Use cookiecutter hooks instead of POST file
+- Refactor templates project name and slug
+
+### Fixed
+- Add coverage report unit test
+
 ## [0.0.5]
 
 ### Added